livekit-agent/requirements.txt
livekit-agents>=0.8.0
livekit-plugins-deepgram>=0.6.0
livekit-plugins-elevenlabs>=0.6.0
livekit-plugins-openai>=0.6.0
openai>=1.12.0
python-dotenv>=1.0.0
supabase>=2.3.0
asyncpg>=0.29.0
aiohttp>=3.9.0
